refactor: move delayed and static scaling to StatefulFloat8Linear (#1439)
float8linear: refactor stateful logic to separate module

Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:
