+++ test/float8/test_base.py
@@ -55,9 +55,11 @@ from torchao.float8.float8_tensor import (
 from torchao.float8.float8_utils import (
     FP8_TYPES,
     compute_error,
+    config_has_stateful_scaling,
     fp8_tensor_statistics,
     tensor_to_scale,
 )
+from torchao.float8.stateful_float8_linear import StatefulFloat8Linear
 from torchao.testing.float8.test_utils import get_test_float8_linear_config
 
 random.seed(0)
@@ -279,10 +281,17 @@ class TestFloat8Linear:
         config: Float8LinearConfig,
         use_ac: bool = False,
     ):
-        m_fp8 = Float8Linear.from_float(
-            copy.deepcopy(m_ref),
-            config,
-        )
+        if config_has_stateful_scaling(config):
+            m_fp8 = StatefulFloat8Linear.from_float(
+                copy.deepcopy(m_ref),
+                config,
+            )
+        else:
+            m_fp8 = Float8Linear.from_float(
+                copy.deepcopy(m_ref),
+                config,
+            )
+
         for _ in range(2):
             if use_ac:
                 y_fp8 = torch.utils.checkpoint.checkpoint(m_fp8, x, use_reentrant=False)
