+++ torchao/float8/float8_linear_utils.py
@@ -15,7 +15,9 @@ from torchao.float8.config import Float8LinearConfig, ScalingType
 from torchao.float8.float8_linear import Float8Linear
 from torchao.float8.float8_utils import (
     amax_history_to_scale_stack,
+    config_has_stateful_scaling,
 )
+from torchao.float8.stateful_float8_linear import StatefulFloat8Linear
 
 log = logging.getLogger(__name__)
 log.addHandler(logging.NullHandler())
@@ -141,10 +143,18 @@ def convert_to_float8_training(
     """
     if config is None:
         config = Float8LinearConfig()
-    from_float = lambda m: Float8Linear.from_float(
-        m,
-        config=config,
-    )
+
+    if config_has_stateful_scaling(config):
+        from_float = lambda m: StatefulFloat8Linear.from_float(
+            m,
+            config=config,
+        )
+    else:
+        from_float = lambda m: Float8Linear.from_float(
+            m,
+            config=config,
+        )
+
     return swap_linear_layers(
         module,
         from_float,
